// 08-Desenvolvimento-Avançado-em-Angular/09-Projeto-Final/front-end/src/app/fornecedor/models/endereco.ts
export interface Endereco {
    id: string;
    logradouro: string;
    numero: string;
    complemento: string;
    bairro: string;
    cep: string;
    cidade: string;
    estado: string;
    fornecedorId: string;
}

export interface CepConsulta{
  cep: string;
  logradouro: string;
  complemento: string;
  bairro: string;
  localidade: string;
  uf: string;
}